Refs #23296 Fix compatibility issues.
