+++ scripts/test/Muon/CMakeLists.txt
@@ -7,7 +7,9 @@ set ( TEST_PY_FILES
    FFTModel_test.py
    MaxEntPresenter_test.py
    MaxEntModel_test.py
+   threading_manager_cancel_test.py
    threading_manager_test.py
+   threading_manager_worker_test.py
    transformWidget_test.py
 )
